+++ sperl_constant_pool_sub.h
@@ -6,11 +6,11 @@
 // Array
 struct SPerl_constant_pool_sub {
   void* native_address;
-  size_t bytecode_base;
-  size_t abs_name_constant_pool_address;
-  size_t operand_stack_max;
-  size_t my_vars_length;
-  size_t args_length;
+  int32_t bytecode_base;
+  int32_t abs_name_constant_pool_address;
+  int32_t operand_stack_max;
+  int32_t my_vars_length;
+  int32_t args_length;
   _Bool is_native;
   _Bool has_return_value;
 };
